+++ Game/src/audiolib/multivoc.h
@@ -88,7 +88,7 @@ int   MV_StartPlayback( void );
 void  MV_StopPlayback( void );
 int   MV_StartRecording( int MixRate, void ( *function )( char *ptr, int length ) );
 void  MV_StopRecord( void );
-int   MV_StartDemandFeedPlayback( void ( *function )( char **ptr, unsigned long *length ),
+int   MV_StartDemandFeedPlayback( void ( *function )( char **ptr, uint32_t *length ),
          int rate, int pitchoffset, int vol, int left, int right,
          int priority, unsigned long callbackval );
 int   MV_PlayRaw( char *ptr, unsigned long length,
